+++ malachite-nz/src/gaussian_integer/logic/mod.rs
@@ -6,16 +6,7 @@
 // Lesser General Public License (LGPL) as published by the Free Software Foundation; either version
 // 3 of the License, or (at your option) any later version. See <https://www.gnu.org/licenses/>.
 
-use malachite_base::test_util::runner::Runner;
-
-pub(crate) fn register(runner: &mut Runner) {
-    arithmetic::register(runner);
-    comparison::register(runner);
-    conversion::register(runner);
-    logic::register(runner);
-}
-
-mod arithmetic;
-mod comparison;
-mod conversion;
-mod logic;
+/// An implementation of [`SignificantBits`](malachite_base::num::logic::traits::SignificantBits), a
+/// trait for determining how many significant bits a number has, and the related
+/// [`max_significant_bits`](crate::gaussian_integer::GaussianInteger::max_significant_bits).
+pub mod significant_bits;