+++ malachite-q/src/bin_util/demo_and_bench/gaussian_rational/logic/mod.rs
@@ -9,13 +9,7 @@
 use malachite_base::test_util::runner::Runner;
 
 pub(crate) fn register(runner: &mut Runner) {
-    arithmetic::register(runner);
-    comparison::register(runner);
-    conversion::register(runner);
-    logic::register(runner);
+    significant_bits::register(runner);
 }
 
-mod arithmetic;
-mod comparison;
-mod conversion;
-mod logic;
+mod significant_bits;