+++ malachite-q/src/bin_util/demo_and_bench/gaussian_rational/mod.rs
@@ -12,8 +12,10 @@ pub(crate) fn register(runner: &mut Runner) {
     arithmetic::register(runner);
     comparison::register(runner);
     conversion::register(runner);
+    logic::register(runner);
 }
 
 mod arithmetic;
 mod comparison;
 mod conversion;
+mod logic;