+++ malachite-q/src/gaussian_rational/mod.rs
@@ -19,6 +19,8 @@ pub mod comparison;
 pub mod conversion;
 /// Iterators that generate [`GaussianRational`]s without repetition.
 pub mod exhaustive;
+/// Traits for logic and bit manipulation.
+pub mod logic;
 #[cfg(feature = "random")]
 /// Iterators that generate [`GaussianRational`]s randomly.
 pub mod random;